requirements.txt
argmaxtools>=0.1.13
torch
safetensors
mlx
jaxtyping
transformers
pillow
sentencepiece
